+++ packages/doc-gen/src/remark-ts2js.ts
@@ -1,8 +1,6 @@
 import type { Transformer } from 'unified';
 import type { Root } from 'mdast';
 import { visit } from 'unist-util-visit';
-import { defineMdastPlugin } from 'satteri';
-import { fileURLToPath } from 'node:url';
 import {
   generateCodeBlockTabs,
   parseCodeBlockAttributes,
@@ -113,68 +111,4 @@ export function remarkTypeScriptToJavaScript({
 
     await Promise.all(tasks);
   };
-}
-
-export function remarkTypeScriptToJavaScriptSatteri(options: TypeScriptToJavaScriptOptions = {}) {
-  const { persist = false, defaultValue = 'ts', disableTrigger = false } = options;
-
-  return defineMdastPlugin({
-    name: 'remark-typescript-to-javascript',
-    async code(node, ctx) {
-      const lang = node.lang;
-      if (lang !== 'ts' && lang !== 'tsx') return;
-
-      const meta = parseCodeBlockAttributes(node.meta ?? '', ['ts2js']);
-      if (!disableTrigger && !('ts2js' in meta.attributes)) return;
-
-      const filePath = ctx.fileURL ? fileURLToPath(ctx.fileURL) : 'test';
-      const oxc = await import('oxc-transform');
-      const result = await oxc.transform(`${filePath}.${lang}`, node.value, {
-        sourcemap: false,
-        jsx: 'preserve',
-      });
-
-      ctx.replaceNode(
-        node,
-        generateCodeBlockTabs({
-          persist,
-          defaultValue,
-          triggers: [
-            {
-              value: 'ts',
-              children: [{ type: 'text', value: 'TypeScript' }],
-            },
-            {
-              value: 'js',
-              children: [{ type: 'text', value: 'JavaScript' }],
-            },
-          ],
-          tabs: [
-            {
-              value: 'ts',
-              children: [
-                {
-                  type: 'code',
-                  lang: node.lang,
-                  meta: meta.rest,
-                  value: node.value,
-                },
-              ],
-            },
-            {
-              value: 'js',
-              children: [
-                {
-                  type: 'code',
-                  lang: lang === 'tsx' ? 'jsx' : 'js',
-                  meta: meta.attributes.ts2js ?? meta.rest,
-                  value: result.code,
-                },
-              ],
-            },
-          ],
-        }),
-      );
-    },
-  });
 }
