+++ packages/doc-gen/tsdown.config.ts
@@ -8,6 +8,5 @@ export default defineConfig({
   entry: ['src/index.ts', 'src/remark-ts2js.ts'],
   deps: {
     onlyBundle: [],
-    neverBundle: ['satteri'],
   },
 });
