+++ packages/satteri/src/index.ts
@@ -38,6 +38,7 @@ export type { StructuredData } from '@/remark-structure';
 export { remarkImage, type RemarkImageOptions } from '@/remark-image';
 export { remarkCodeTab, type RemarkCodeTabOptions } from '@/remark-code-tab';
 export { remarkNpm, type RemarkNpmOptions } from '@/remark-npm';
+export { remarkTs2js, type RemarkTs2jsOptions } from '@/remark-ts2js';
 export { remarkBlockId, type RemarkBlockIdOptions } from '@/remark-block-id';
 export { remarkAdmonition, type RemarkAdmonitionOptions } from '@/remark-admonition';
 export {
